Roadmap/voice-agent-pro/stt/stt2/requirements.txt
deepgram-sdk==2.12.0
python-dotenv
jiwer
